+++ CHANGELOG.md
@@ -25,6 +25,7 @@ auto-detected): includes spatial datasets and extended to mcd & tiff
 - Using prev or next ROI before a key listener has been initiated blocks the callback
 - Allow y-axis overflow for multiple gating selection dropdown
 - `pymongo` client uses `certifi.where()` for SSL certificate to be compatible with `pyinstaller`
+- `None` pixel intensity lower and upper bounds for channels on slider checked separately
 
 ### Added